backend/requirements.txt
fastapi>=0.121.2
langchain-community>=0.4.1
langchain-core>=1.0.3
langchain-openai>=0.0.1
numpy>=2.3.4
uvicorn>=0.38.0
python-dotenv>=1.0.0
faiss-cpu>=1.7.4